requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy
httpx
aiocache
PyJWT
Werkzeug
python-dotenv